prevent build making a copy into the wrong opnsense-plugins branch

+++ extras/build-tools/build-at-remote.sh
@@ -6,6 +6,7 @@ remote_host=${1}
 remote_user=root
 
 opnsense_plugins_repo_path=$(realpath $(dirname $(realpath $0))/../../../opnsense-plugins)
+opnsense_plugins_branch='autossh'
 autossh_plugin_repo_path=$(realpath $(dirname $(realpath $0))/../../)
 autossh_opnsense_plugins_subpath='net/autossh'
 remote_base_path='/root/opnsense-plugins'
@@ -17,12 +18,24 @@ fi
 
 # echo some helpful output
 echo ""
-echo "remote:                            ${remote_user}@${remote_host}"
-echo "remote_base_path:                  ${remote_base_path}"
-echo "opnsense_plugins_repo_path:        ${opnsense_plugins_repo_path}"
-echo "autossh_plugin_repo_path:          ${autossh_plugin_repo_path}"
-echo "autossh_opnsense_plugins_subpath:  ${autossh_opnsense_plugins_subpath}"
+echo "remote:                               ${remote_user}@${remote_host}"
+echo "remote_base_path:                     ${remote_base_path}"
+echo "opnsense_plugins_repo_path:           ${opnsense_plugins_repo_path}"
+echo "opnsense_plugins_branch:              ${opnsense_plugins_branch}"
+echo "autossh_plugin_repo_path:             ${autossh_plugin_repo_path}"
+echo "autossh_opnsense_plugins_subpath:     ${autossh_opnsense_plugins_subpath}"
 echo ""
+
+# make sure we are on the correct branch
+current_plugins_branch=$(cd ${opnsense_plugins_repo_path}; git rev-parse --abbrev-ref HEAD)
+if [ ${current_plugins_branch} != ${opnsense_plugins_branch} ]; then
+    echo "OPNsense Plugins is not on the correct branch!"
+    echo " - path: ${opnsense_plugins_repo_path}"
+    echo " - current: ${current_plugins_branch}"
+    echo " - expected: ${opnsense_plugins_branch}"
+    echo ""
+    exit 1
+fi
 
 # copy into place the material to be injected into the repo
 rm -Rf ${opnsense_plugins_repo_path}/${autossh_opnsense_plugins_subpath}
